+++ Assets/_Scripts.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 36a7401d1ceb57449bce8d7d337d069c
+folderAsset: yes
+timeCreated: 1480351205
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
