+++ Assets/_Scripts/GSO.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: baed8bdee877e5c40a6b3bff08a2f405
+timeCreated: 1480349749
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
