+++ Assets/_Scripts/InitScript.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: bb357baa4de7fa144b4d0b32296fb8c5
+timeCreated: 1480349749
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
