+++ Assets/_Scripts/ObstacleTrigger.cs
