+++ Assets/_Scripts/ObstacleTrigger.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 34574236a997b034fa50556e2e4a304d
+timeCreated: 1480349749
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
