+++ Assets/_Scripts/PlayerMove.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 50198187e8349e548a862a04aa1789c4
+timeCreated: 1480349749
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
