+++ ProjectSettings/ProjectVersion.txt
@@ -1,0 +1,2 @@
+m_EditorVersion: 5.4.2f2
+m_StandardAssetsVersion: 0
