+++ README.md
@@ -1,0 +1,5 @@
+# Glowworm Swarm Optimization in Unity
+
+- Original Project by kanishkaganguly
+- Using Unity 5.4.2f2
+- Based on the paper [Glowworm swarm optimization for simultaneous capture of multiple local optima of multimodal functions](./GSO_Paper.pdf)
